requirements.txt
streamlit>=1.29.0
openai>=1.3.0
pinecone>=2.2.4
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
tiktoken>=0.5.0
pyperclip>=1.8.2
# Optional additions:
pypdf2>=3.0.0
langchain>=0.0.350